requirements.txt
fastapi
uvicorn
praw
yfinance
pandas
numpy
scikit-learn
nltk
python-dotenv
httpx
